requirements.txt
pyalsaaudio
aubio
python-rtmidi
